requirements.txt
asgiref==3.7.2
Brotli==1.1.0
cffi==1.16.0
cssselect2==0.7.0
Django==4.2.9
fonttools==4.49.0
html5lib==1.1
pillow==10.2.0
pycparser==2.21
pydyf==0.9.0
pyphen==0.14.0
python-barcode==0.15.1
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.9.0
weasyprint==61.1
webencodings==0.5.1
zopfli==0.2.3
